+++ modules/backend/ServiceProvider.php
@@ -7,6 +7,8 @@ use BackendMenu;
 use BackendAuth;
 use Backend\Classes\WidgetManager;
 use October\Rain\Support\ModuleServiceProvider;
+use System\Models\EmailTemplate;
+use System\Classes\SettingsManager;
 
 class ServiceProvider extends ModuleServiceProvider
 {
@@ -62,12 +64,38 @@ class ServiceProvider extends ModuleServiceProvider
         });
 
         /*
+         * Register settings
+         */
+        SettingsManager::instance()->registerCallback(function($manager){
+            $manager->registerSettingItems('October.Backend', [
+                'editor' => [
+                    'label'       => 'backend::lang.editor.menu_label',
+                    'description' => 'backend::lang.editor.menu_description',
+                    'category'    => 'System',
+                    'icon'        => 'icon-code',
+                    'class'       => 'Backend\Models\EditorSettings',
+                    'sort'        => 200
+                ],
+            ]);
+        });
+
+        /*
          * Register permissions
          */
         BackendAuth::registerCallback(function($manager) {
             $manager->registerPermissions('October.Backend', [
                 'backend.access_dashboard' => ['label' => 'View the dashboard', 'tab' => 'System'],
                 'backend.manage_users'     => ['label' => 'Manage other administrators', 'tab' => 'System'],
+            ]);
+        });
+
+        /*
+         * Register email templates
+         */
+        EmailTemplate::registerCallback(function($template){
+            $template->registerEmailTemplates([
+                'backend::emails.invite' => 'Invitation for newly created administrators.',
+                'backend::emails.restore' => 'Password reset instructions for backend-end administrators.',
             ]);
         });
     }