+++ modules/backend/behaviors/RelationController.php
@@ -116,7 +116,7 @@ class RelationController extends ControllerBehavior
     {
         parent::__construct($controller);
 
-        $this->addJs('js/october.relation.js');
+        $this->addJs('js/october.relation.js', 'core');
 
         /*
          * Build configuration
@@ -304,6 +304,7 @@ class RelationController extends ControllerBehavior
         $this->vars['relationManageId'] = $this->manageId;
         $this->vars['relationLabel'] = $this->config->label ?: $this->field;
         $this->vars['relationField'] = $this->field;
+        $this->vars['relationType'] = $this->relationType;
         $this->vars['relationViewWidget'] = $this->viewWidget;
         $this->vars['relationManageWidget'] = $this->manageWidget;
         $this->vars['relationToolbarWidget'] = $this->toolbarWidget;
@@ -421,7 +422,18 @@ class RelationController extends ControllerBehavior
 
     public function onRelationManageDelete()
     {
-        // @todo
+        $this->beforeAjax();
+
+        if (($checkedIds = post('checked')) && is_array($checkedIds)) {
+            foreach ($checkedIds as $relationId) {
+                if (!$obj = $this->relationObject->find($relationId))
+                    continue;
+
+                $obj->delete();
+            }
+        }
+
+        return ['#'.$this->relationGetId('view') => $this->relationRenderView()];
     }
 
     public function onRelationManageCreate()
@@ -430,6 +442,16 @@ class RelationController extends ControllerBehavior
 
         $saveData = $this->manageWidget->getSaveData();
         $this->relationObject->create($saveData);
+
+        return ['#'.$this->relationGetId('view') => $this->relationRenderView()];
+    }
+
+    public function onRelationManageUpdate()
+    {
+        $this->beforeAjax();
+
+        $saveData = $this->manageWidget->getSaveData();
+        $this->relationObject->find($this->manageId)->save($saveData);
 
         return ['#'.$this->relationGetId('view') => $this->relationRenderView()];
     }